Consolidate arch* files
