+++ ddprof-lib/src/main/cpp/livenessTracker.h
@@ -6,7 +6,7 @@
 #ifndef _LIVENESSTRACKER_H
 #define _LIVENESSTRACKER_H
 
-#include "arch_dd.h"
+#include "arch.h"
 #include "context.h"
 #include "engine.h"
 #include "event.h"
