+++ ddprof-lib/src/main/cpp/perfEvents.h
@@ -19,7 +19,7 @@
 
 #ifdef __linux__
 
-#include "arch_dd.h"
+#include "arch.h"
 #include "arguments.h"
 #include "engine.h"
 #include <signal.h>
