+++ ddprof-lib/src/main/cpp/threadFilter.h
@@ -22,7 +22,7 @@
 #include <cstdint>
 #include <memory>
 
-#include "arch_dd.h"
+#include "arch.h"
 
 class ThreadFilter {
 public:
